+++ CHANGELOG.md
@@ -7,6 +7,7 @@ InfrastructureModels.jl Change Log
 ### v0.0.2
 - Added dict summary function
 - Added update_data! function
+- Added replicate function
 - Added basic documentation
 
 ### v0.0.1